# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WA App API",
        "description": "API per gestione post, media e categorie",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server locale"
        },
        {
            "url": "https://tuosito.com",
            "description": "Server produzione"
        }
    ],
    "paths": {
        "/api/category-types": {
            "post": {
                "tags": [
                    "CategoryTypes"
                ],
                "summary": "Crea un nuovo tipo di categoria",
                "description": "Endpoint per creare un nuovo tipo di categoria",
                "operationId": "981c8bae02c7071958be61c0a8681f1e",
                "requestBody": {
                    "description": "Dati del tipo categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tipo A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo categoria creato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validazione fallita"
                    }
                }
            }
        },
        "/api/media": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Lista di tutti i media",
                "description": "Restituisce una lista paginata di media con supporto a filtri e ordinamenti",
                "operationId": "8edb939568a1e19d1fbef2627d7d4ad2",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filtra per nome (ricerca parziale)",
                        "schema": {
                            "type": "string",
                            "example": "logo"
                        }
                    },
                    {
                        "name": "filter[file_name]",
                        "in": "query",
                        "description": "Filtra per nome file (ricerca parziale)",
                        "schema": {
                            "type": "string",
                            "example": "logo.png"
                        }
                    },
                    {
                        "name": "filter[mime_type]",
                        "in": "query",
                        "description": "Filtra per tipo MIME (match esatto)",
                        "schema": {
                            "type": "string",
                            "example": "image/png"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordinamento: created_at, name, file_name, size (usa - per discendente)",
                        "schema": {
                            "type": "string",
                            "example": "-created_at"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero pagina (default 1)",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista media",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "file_name": {
                                                        "type": "string"
                                                    },
                                                    "mime_type": {
                                                        "type": "string"
                                                    },
                                                    "size": {
                                                        "type": "integer"
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/media/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Dettaglio di un singolo media",
                "description": "Restituisce i dati completi di un media specifico",
                "operationId": "b690ac08db13263dbfd3d13da62bf081",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del media",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dettaglio media",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "file_name": {
                                            "type": "string"
                                        },
                                        "file_path": {
                                            "type": "string"
                                        },
                                        "mime_type": {
                                            "type": "string"
                                        },
                                        "size": {
                                            "type": "integer"
                                        },
                                        "url": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Media non trovato"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista di tutti i post",
                "description": "Restituisce una lista paginata di post con supporto a filtri e ordinamenti",
                "operationId": "9ece41c64501f4e238ad74f40d0197f3",
                "parameters": [
                    {
                        "name": "filter[title]",
                        "in": "query",
                        "description": "Filtra per titolo (ricerca parziale)",
                        "schema": {
                            "type": "string",
                            "example": "news"
                        }
                    },
                    {
                        "name": "filter[excerpt]",
                        "in": "query",
                        "description": "Filtra per estratto (ricerca parziale)",
                        "schema": {
                            "type": "string",
                            "example": "aggiornamento"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordinamento: created_at, title (usa - per discendente)",
                        "schema": {
                            "type": "string",
                            "example": "-created_at"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero pagina (default 1)",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "excerpt": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Dettaglio di un singolo post",
                "description": "Restituisce i dati completi di un post specifico",
                "operationId": "d0cd62ddd083b5075be434ba0d9a17d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dettaglio post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        },
                                        "excerpt": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post non trovato"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "CategoryTypes",
            "description": "CategoryTypes"
        },
        {
            "name": "Media",
            "description": "Media"
        },
        {
            "name": "Posts",
            "description": "Posts"
        }
    ]
}